move internal/kafka to pkg/kafka
