+++ cmd/otelcontribcol/builder-config.yaml
@@ -328,7 +328,7 @@ replaces:
   - github.com/open-telemetry/opentelemetry-collector-contrib/exporter/azuredataexplorerexporter => ../../exporter/azuredataexplorerexporter
   - github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8sconfig => ../../internal/k8sconfig
   - github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8stest => ../../internal/k8stest
-  - github.com/open-telemetry/opentelemetry-collector-contrib/internal/kafka => ../../internal/kafka
+  - github.com/open-telemetry/opentelemetry-collector-contrib/pkg/kafka => ../../internal/kafka
   - github.com/open-telemetry/opentelemetry-collector-contrib/receiver/carbonreceiver => ../../receiver/carbonreceiver
   - github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter => ../../exporter/splunkhecexporter
   - github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusexporter => ../../exporter/prometheusexporter
