+++ pkg/kafka/authentication_test.go
