+++ pkg/kafka/awsmsk/iam_scram_client_test.go
